+++ example_codes/keyboard_vs_keyboard_example.py
@@ -1,7 +1,6 @@
 import retro
 from pyglet.window import key
 import numpy as np
-import random
 
 from pathlib import Path
 from time import sleep
@@ -17,28 +16,27 @@ env = retro.make(env_name,
     state=str(state_path),
     players=2)
 
-class Agent(object):
-    def __init__(self, action_space):
-        self.action_space = action_space
-    
-    #return a random move
-    def act(self, observation, reward, done):
-        i =  int(2 ** (random.random() * 12 ))
-        return  f'{i|2048:012b}' 
-
-PLAYER1 = 0
+PLAYER1, PLAYER2 = 0, 12
 LOW_KICK, PUNCH, BLOCK, JUMP, SQUAT, MOVE_LEFT, MOVE_RIGHT, HIGH_KICK = 0, 1, 3, 4, 5, 6, 7, 8
 
-pushed_keys = list('000000000000') 
+pushed_keys = list('000000000000'+'000000000000')
 keymap = {
+          key.UP: PLAYER2+JUMP,
+          key.DOWN: PLAYER2+SQUAT,
+          key.LEFT: PLAYER2+MOVE_LEFT,
+          key.RIGHT: PLAYER2+MOVE_RIGHT,
+          key.MINUS: PLAYER2+BLOCK,
+          key.RCTRL: PLAYER2+LOW_KICK,
+          key.PERIOD: PLAYER2+HIGH_KICK,
+          key.RSHIFT: PLAYER2+PUNCH,
           key.W: PLAYER1+JUMP,
           key.S: PLAYER1+SQUAT,
           key.A: PLAYER1+MOVE_LEFT,
           key.D: PLAYER1+MOVE_RIGHT,
-          key.LEFT: PLAYER1+BLOCK,
-          key.DOWN: PLAYER1+LOW_KICK,
-          key.UP: PLAYER1+HIGH_KICK,
-          key.RIGHT: PLAYER1+PUNCH
+          key.Q: PLAYER1+BLOCK,
+          key.LCTRL: PLAYER1+LOW_KICK,
+          key.LSHIFT: PLAYER1+HIGH_KICK,
+          key.E: PLAYER1+PUNCH
           }
 
 def key_press(pressed_key, modifier):
@@ -57,13 +55,10 @@ state = env.reset()
 env.render()
 env.viewer.window.on_key_press = key_press
 env.viewer.window.on_key_release = key_release
-agent = Agent(env.action_space)
-action = f'{0|2048:012b}'
-ob, reward, done, _ = env.step( action + "".join(pushed_keys) )
+
 while True:
     for i in range(100000):
-        action = agent.act(ob, reward, done)
-        ob, reward, done, _ = env.step("".join(pushed_keys) + action )
+        observation, reward, done, info = env.step("".join(pushed_keys))
         env.render()
-        sleep(1/90)
+        sleep(0.001)
     env.reset()
